requirements.txt
torch
torchvision
datasets
scikit-learn
pillow
matplotlib
wandb
tqdm
